Probably working websocket to ble communication for movement
