+++ server/models.py
@@ -21,10 +21,12 @@ class Car:
         self.ble_address = ble_address  # BLE MAC address
         self.battery_level = 100  # Battery level percentage (0-100)
         self.move = "stopped"     # Current movement: stopped, forward, backward, left, right
+        self.x = 0                # Steering direction (-100 to 100, left to right)
         self.boost = False        # Whether boost is currently active
         self.boost_value = 100    # Boost fuel level (0-100)
         self.connected = False    # Whether the car is connected via Bluetooth
         self.last_seen = None     # Timestamp of last BLE discovery
+        self.websocket_id = None  # WebSocket connection identifier that controls this car
         
     def update_status(self, **kwargs):
         """
@@ -51,14 +53,18 @@ class Car:
             "ble_address": self.ble_address,
             "battery_level": self.battery_level,
             "move": self.move,
+            "x": self.x,
             "boost": str(self.boost).lower(),
             "boost_value": self.boost_value,
             "connected": self.connected,
-            "last_seen": self.last_seen.isoformat() if self.last_seen else None
+            "last_seen": self.last_seen.isoformat() if self.last_seen else None,
+            "websocket_id": self.websocket_id,
+            "selected": self.websocket_id is not None
         }
     
     def __str__(self):
-        return f"Car {self.car_id} ({self.name}) - BLE: {self.ble_name} - Battery: {self.battery_level}%, Move: {self.move}, Boost: {self.boost}"
+        selected_status = f" [Selected by {self.websocket_id}]" if self.websocket_id else " [Available]"
+        return f"Car {self.car_id} ({self.name}) - BLE: {self.ble_name} - Battery: {self.battery_level}%, Move: {self.move}, Boost: {self.boost}{selected_status}"
     
     def __repr__(self):
         return f"Car(car_id={self.car_id}, name='{self.name}', ble_name='{self.ble_name}')"
@@ -249,3 +255,90 @@ class CarManager:
             mac_part = ble_name.replace("RL-CAR-", "")
             return f"Rocket League Car ({mac_part[-8:]})"  # Use last 8 chars of MAC
         return f"Unknown Car ({ble_name})"
+    
+    def select_car(self, car_id, websocket_id):
+        """
+        Select/assign a car to a websocket connection.
+        
+        Args:
+            car_id (int): ID of the car to select
+            websocket_id (str): Unique identifier for the websocket connection
+            
+        Returns:
+            tuple: (success: bool, message: str, car: Car or None)
+        """
+        car = self.get_car(car_id)
+        if not car:
+            return False, f"Car {car_id} not found", None
+        
+        if car.websocket_id is not None:
+            if car.websocket_id == websocket_id:
+                return True, f"Car {car_id} is already selected by this client", car
+            else:
+                return False, f"Car {car_id} is already selected by another client", None
+        
+        car.websocket_id = websocket_id
+        return True, f"Car {car_id} successfully selected", car
+    
+    def free_car(self, car_id, websocket_id=None):
+        """
+        Free/release a car from its websocket connection.
+        
+        Args:
+            car_id (int): ID of the car to free
+            websocket_id (str, optional): WebSocket ID that should own the car (for verification)
+            
+        Returns:
+            tuple: (success: bool, message: str)
+        """
+        car = self.get_car(car_id)
+        if not car:
+            return False, f"Car {car_id} not found"
+        
+        if car.websocket_id is None:
+            return True, f"Car {car_id} is already free"
+        
+        # If websocket_id is provided, verify ownership
+        if websocket_id is not None and car.websocket_id != websocket_id:
+            return False, f"Car {car_id} is not selected by this client"
+        
+        car.websocket_id = None
+        return True, f"Car {car_id} has been freed"
+    
+    def free_cars_by_websocket(self, websocket_id):
+        """
+        Free all cars assigned to a specific websocket connection.
+        
+        Args:
+            websocket_id (str): WebSocket connection identifier
+            
+        Returns:
+            list: List of car IDs that were freed
+        """
+        freed_cars = []
+        for car in self.cars.values():
+            if car.websocket_id == websocket_id:
+                car.websocket_id = None
+                freed_cars.append(car.car_id)
+        return freed_cars
+    
+    def get_free_cars(self):
+        """
+        Get all cars that are not currently selected by any websocket.
+        
+        Returns:
+            list: List of Car objects that are available
+        """
+        return [car for car in self.cars.values() if car.websocket_id is None]
+    
+    def get_cars_by_websocket(self, websocket_id):
+        """
+        Get all cars assigned to a specific websocket connection.
+        
+        Args:
+            websocket_id (str): WebSocket connection identifier
+            
+        Returns:
+            list: List of Car objects assigned to the websocket
+        """
+        return [car for car in self.cars.values() if car.websocket_id == websocket_id]
